refactor: 최종 결과 문자열 관련기능 클래스 분리
- 최종 결과 문자열에 관련된 기능들을 ResultMessage 클래스로 분리함.
